Update build process; move files and folders
The Choice block needs to access the 'interactive-fiction-engine'
and 'dynamic-choice' scripts inside of the editor, but we also
need to instantiate logic using those scripts on the frontend.

I moved around files and logic in order to accomplish this.

As of this commit, we now have a build process for each scenario:
npm run start:choice
npm run start:js

To improve this, we could next consolidate both processes under
one command.